requirements.txt
selenium==4.8.1
beautifulsoup4==4.11.2
SQLAlchemy==2.0.15
python-dotenv==1.0.0
nltk==3.8.1